requirements.txt
flask~=3.0.3
PyYAML~=6.0.1